Moving AffiliateCredits to utilities and removing index file.
