+++ resources/assets/components/utilities/AffiliateCredits/AffiliateCredits.js
